requirements.txt
# Merger Book MVP Requirements

# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-authenticator>=0.2.3

# Database
sqlite3

# Environment and configuration
python-dotenv>=1.0.0

# AI and NLP
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.0

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
pdf2image>=1.16.3
Pillow>=10.0.0

# Financial data APIs
polygon-api-client>=1.12.0
yfinance>=0.2.18
requests>=2.31.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3
hashlib
json
datetime
typing
os
sys

